refactor: extract expand_managed_services as standalone function
