+++ src/airflow/managed_services.rs
@@ -1,4 +1,5 @@
 pub mod astronomer;
 pub mod composer;
 pub mod conveyor;
+pub mod expand;
 pub mod mwaa;
